re: #3776
add tiger installer for nonX86
move the install-tiger.sh to both distros

+++ nimbix_desktop/install-tiger.sh
@@ -5,7 +5,12 @@ ARCH=$(arch)
 
 if [ "$ARCH" != "x86_64" ]; then
     #build_and_install_tiger
-    yum -y install tigervnc-server
+    if [[ -f /etc/redhat-release ]]; then
+        yum -y install tigervnc-server
+    else
+        apt-get -y update
+        apt-get -y install tigervnc-standalone-server
+    fi
 else
     # Install the cached tarball
     sudo tar -C / -xzf  /usr/local/lib/nimbix_desktop/tigervnc-$VERSION.$ARCH.tar.gz --strip-components=1
